django_want/requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
Django==3.1
django-cors-headers==3.7.0
django-filter==2.3.0
django-oauth-toolkit==1.3.2
djangorestframework==3.11.1
flake8==3.8.3
gunicorn==20.0.4
idna==2.10
Markdown==3.2.2
mccabe==0.6.1
oauthlib==3.1.0
psycopg2==2.8.5
pycodestyle==2.6.0
pyflakes==2.2.0
python-dotenv==0.15.0
pytz==2020.1
qiniu==7.2.10
requests==2.24.0
sqlparse==0.3.1
urllib3==1.25.10
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
coverage==6.1.2
Django==3.1
django-cors-headers==3.7.0
django-filter==2.3.0
django-oauth-toolkit==1.3.2
djangorestframework==3.11.1
flake8==3.8.3
gunicorn==20.0.4
idna==2.10
Markdown==3.2.2
mccabe==0.6.1
oauthlib==3.1.0
psycopg2==2.8.5
pycodestyle==2.6.0
pyflakes==2.2.0
python-dotenv==0.15.0
pytz==2020.1
qiniu==7.2.10
requests==2.24.0
sqlparse==0.3.1
urllib3==1.25.10
